Add attribute query methods to base class to inherit from
