+++ lib/jsonb_accessor/adapters.rb
@@ -6,4 +6,3 @@ end
 require_relative "adapters/base_adapter"
 require_relative "adapters/postgresql_adapter"
 require_relative "adapters/sqlite_adapter"
-
