+++ lib/jsonb_accessor/adapters/base_adapter.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+# Loaded but not included. It is meant for inheriting from
+require_relative "base_adapter/attribute_query_methods"
+
 module JsonbAccessor::Adapters::BaseAdapter
   module_function
