+++ lib/jsonb_accessor/adapters/sqlite_adapter/attribute_query_methods.rb
@@ -3,27 +3,10 @@
 module JsonbAccessor
   module Adapters
     module SqliteAdapter
-      class AttributeQueryMethods
-        def initialize(klass)
-          @klass = klass
-        end
+      class AttributeQueryMethods < JsonbAccessor::Adapters::BaseAdapter::AttributeQueryMethods
+        private
 
-        def define(store_key_mapping_method_name, jsonb_attribute)
-          return if klass.superclass.respond_to? store_key_mapping_method_name
-
-          # <jsonb_attribute>_where scope
-          klass.define_singleton_method "#{jsonb_attribute}_where" do |attributes|
-            store_key_attributes = ::JsonbAccessor::Helpers.convert_keys_to_store_keys(attributes, all.model.public_send(store_key_mapping_method_name))
-            jsonb_where(jsonb_attribute, store_key_attributes)
-          end
-
-          # <jsonb_attribute>_where_not scope
-          klass.define_singleton_method "#{jsonb_attribute}_where_not" do |attributes|
-            store_key_attributes = ::JsonbAccessor::Helpers.convert_keys_to_store_keys(attributes, all.model.public_send(store_key_mapping_method_name))
-            jsonb_where_not(jsonb_attribute, store_key_attributes)
-          end
-
-          # <jsonb_attribute>_order scope
+        def define_order(store_key_mapping_method_name, jsonb_attribute)
           klass.define_singleton_method "#{jsonb_attribute}_order" do |*args|
             ordering_options = args.extract_options!
             order_by_defaults = args.each_with_object({}) { |attribute, config| config[attribute] = :asc }
@@ -37,10 +20,6 @@ module JsonbAccessor
             end
           end
         end
-
-        private
-
-        attr_reader :klass
       end
     end
   end
